Add repeat checker phases to the compiler
The aim of this commit is to make the compiler to repeat the checker
compilation phases (from namer to refchecks) whenever a phase which is
after refchecks and before icode exits.

That way we can verify that the entire AST is consistent with the
SymbolTable and pickled data.

This is still a work in progress and does not really work properly.